CinemaTicketAutomation
I wrote this project with java language. The project allows the film information to be displayed to the customer by pulling it from the database. After the customer selects the film, the sessions of the related film are shown again by pulling from the database. After the customer selects the session, he/she enters the number of tickets he/she wants to buy and the ticket stock is updated.